eth: moved eth.Config to eth/protocols/eth
